convert the scanning routines to C
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@2750 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/tex/packaging.c
@@ -767,7 +767,7 @@ halfword hpack(halfword p, scaled w, int m)
         glue_order(r) = o;
         glue_sign(r) = shrinking;
         if (total_shrink[o] != 0) {
-            glue_set(r) = unfloat((double) (-x) / total_shrink[o]);
+            glue_set(r) = unfloat((double) (-x) / (double)total_shrink[o]);
         } else {
             glue_sign(r) = normal;
             set_glue_ratio_zero(glue_set(r));   /* there's nothing to shrink */
